add lock_owner

+++ locky_mysql.go
@@ -8,43 +8,8 @@ import (
 	"time"
 )
 
-type Opt struct {
-	Db         *sql.DB
-	Table      string
-	Ctx        context.Context
-	AutoCreate bool
-}
-
-func (o *Opt) Default() {
-	if len(o.Table) == 0 {
-		o.Table = DefaultTable
-	}
-
-	if o.Ctx == nil {
-		o.Ctx = context.Background()
-	}
-}
-
-func (o *Opt) Validate() error {
-	if o.Db == nil {
-		return errors.New("invalid db")
-	}
-
-	if len(o.Table) > 255 {
-		return errors.New("table name too long(>255)")
-	}
-
-	if err := o.Db.PingContext(o.Ctx); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 type MysqlDistributedLock struct {
-	db                   *sql.DB
-	table                string
-	ctx                  context.Context
+	*Opt
 	lockStat, unlockStat *sql.Stmt
 }
 
@@ -53,22 +18,22 @@ const (
 	CreateDDL    = "CREATE TABLE if not exists %s" +
 		"( " +
 		"    `lock_name`     VARCHAR(255)    NOT NULL, " +
+		"    `lock_owner`     VARCHAR(255)    NOT NULL, " +
 		"    `lock_timestamp` BIGINT UNSIGNED NOT NULL, " +
 		"    `lock_ttl`       INT UNSIGNED    NOT NULL, " +
 		"    PRIMARY KEY (`lock_name`) " +
 		") ENGINE = InnoDB " +
 		"  DEFAULT CHARSET = utf8;"
-	QueryLock = "INSERT INTO %s (`lock_name`, `lock_timestamp`, `lock_ttl`)" +
-		"VALUES (?, UNIX_TIMESTAMP(), ?)" +
+	QueryLock = "INSERT INTO %s (`lock_name`, `lock_owner`, `lock_timestamp`, `lock_ttl`)" +
+		"VALUES (?, ?, UNIX_TIMESTAMP(), ?)" +
 		"ON DUPLICATE KEY UPDATE" +
 		"`lock_timestamp` = IF(UNIX_TIMESTAMP() - `lock_timestamp` > `lock_ttl`, VALUES(`lock_timestamp`), `lock_timestamp`)," +
 		"`lock_ttl` = IF(UNIX_TIMESTAMP() - `lock_timestamp` > `lock_ttl`, VALUES(`lock_ttl`), `lock_ttl`);"
-	QueryUnlock = "DELETE FROM %s WHERE `lock_name` = ?;"
+	QueryUnlock = "DELETE FROM %s WHERE `lock_name` = ? and `lock_owner` = ?;"
 )
 
-func NewMysqlDistributedLock(opt Opt) (DistributedLock, error) {
-	opt.Default()
-	if err := opt.Validate(); err != nil {
+func NewMysqlDistributedLock(opt Opt) (*MysqlDistributedLock, error) {
+	if err := opt.DefaultAndValidate(); err != nil {
 		return nil, err
 	}
 
@@ -92,21 +57,14 @@ func NewMysqlDistributedLock(opt Opt) (DistributedLock, error) {
 	}
 
 	return &MysqlDistributedLock{
-		db:         db,
-		table:      table,
-		ctx:        ctx,
+		Opt:        &opt,
 		lockStat:   lockStat,
 		unlockStat: unlockStat,
 	}, nil
 }
 
-func autoCreate(ctx context.Context, db *sql.DB, table string) error {
-	_, err := db.ExecContext(ctx, fmt.Sprintf(CreateDDL, table))
-	return err
-}
-
-func (l *MysqlDistributedLock) Lock(owner string, ttl time.Duration) (bool, error) {
-	if err := l.validateOwner(owner); err != nil {
+func (l *MysqlDistributedLock) Lock(name string, ttl time.Duration) (bool, error) {
+	if err := l.validateName(name); err != nil {
 		return false, err
 	}
 
@@ -114,7 +72,7 @@ func (l *MysqlDistributedLock) Lock(owner string, ttl time.Duration) (bool, erro
 		return false, err
 	}
 
-	res, err := l.lockStat.Exec(owner, ttl.Seconds())
+	res, err := l.lockStat.Exec(name, l.Owner, ttl.Seconds())
 	if err != nil {
 		return false, err
 	}
@@ -131,18 +89,18 @@ func (l *MysqlDistributedLock) Lock(owner string, ttl time.Duration) (bool, erro
 	}
 }
 
-func (l *MysqlDistributedLock) Unlock(owner string) error {
-	if err := l.validateOwner(owner); err != nil {
+func (l *MysqlDistributedLock) Unlock(name string) error {
+	if err := l.validateName(name); err != nil {
 		return err
 	}
 
-	_, err := l.unlockStat.Exec(owner)
+	_, err := l.unlockStat.Exec(name, l.Owner)
 	return err
 }
 
-func (l *MysqlDistributedLock) validateOwner(owner string) error {
-	if len(owner) <= 0 || len(owner) > 255 {
-		return errors.New("owner len must be between 0-255")
+func (l *MysqlDistributedLock) validateName(name string) error {
+	if len(name) <= 0 || len(name) > 255 {
+		return errors.New("name len must be between 0-255")
 	}
 
 	return nil
@@ -154,4 +112,9 @@ func (l *MysqlDistributedLock) validateTTL(ttl time.Duration) error {
 	}
 
 	return nil
+}
+
+func autoCreate(ctx context.Context, db *sql.DB, table string) error {
+	_, err := db.ExecContext(ctx, fmt.Sprintf(CreateDDL, table))
+	return err
 }
